requirements.txt
uagents>=0.20.0
uagents-core>=0.3.0
hyperon>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
pytest>=7.0.0
asyncio>=3.9.0
uuid>=1.30
datetime>=4.3